project/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
importlib-metadata==5.0.0
Markdown==3.4.1
Pillow==9.3.0
PyJWT==2.3.0
pytz==2022.5
sqlparse==0.4.3
zipp==3.9.0
